Add files via upload
Have worked on the Titanic Data-set to carry out Data Analysis using Python programming language. My work and findings includes:

1) Analysis:
a) Using a histogram for the ‘Age’ variable, I have found out the most dominant age group to be in the range 20-30 years.
b) Again using the histogram too for the ‘Fare’ variable, I have found that most of the tickets bought were under 50.
c) More than three-fifths of the total passengers in the crew were males (64.25%) as opposed to 35.75% of females. 
d) Using a scatter plot, no significant relationship between Age and Fair could be seen.
e) Male passengers, on an average, seemed to be older than their female counterparts. This difference was also found to be statistically significant.
f) Female passengers, are more likely to survive than their female counterparts. 

2) Model-fitting:
I have used a logistic regression algorithm on the data-set to predict the survival of a passenger. My model returned an accuracy of 83.68% and an 84% precision.